Move algorithms to separate file
